QskGraphic: Fix scale factors
We need to know the bounding rect of the graphic to know by how much
we can scale our shape, so we need to pass that to the scaling
methods.
E.g. when there are two shapes whose pen sizes overlap the path rect
left and right, we would calculate wrong scale factors without the
grapics' bounding rect.

Resolves #250