tab bar: Take scroll position into account when making item visible
If the tab bar is scrolled, the position we get from calling
mapFromItem() is less than the real position with a scrolling of 0,
because mapFromItem() does not know about our scroll position.

This is usually not noticed when creating a tab bar, because the
scroll position is (0,0). When calling this method on a scrolled
tab bar, the problem becomes visible.